pyproject.toml
[tool.poetry]
name = "django-diploma"
version = "0.1.0"
description = ""
authors = ["Alexander Chekalovets <alex4ekalovets@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.0"
gunicorn = "^21.2.0"
django-debug-toolbar = "^4.1.0"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
